uninstall_hawk.sh - Prior restarting cron make sure that it's init file exists

+++ bin/uninstall_hawk.sh
@@ -4,7 +4,7 @@
 # copyright@1h.com                                              http://1h.com
 # This code is subject to the 1H license. Unauthorized copying is prohibited.
 
-VERSION='0.0.2'
+VERSION='0.0.3'
 
 #. /usr/local/1h/lib/sh/uninstall_funcs.sh
 function check_err() {
@@ -82,20 +82,24 @@ warn_err $? "Could not stop Hawk daemon"
 rm -f /usr/local/1h/var/run/hawk* /usr/local/1h/var/log/hawk*
 warn_err $? "Could not remove pidfile and logfile"
 
-chattr -ai /var/spool/cron/root
-warn_err $? "Could not chattr -ai /var/spool/cron/root"
+if [ -f /var/spool/cron/root ]; then
+	chattr -ai /var/spool/cron/root
+	warn_err $? "Could not chattr -ai /var/spool/cron/root"
 
-sed -i '/usr\/local\/1h\/bin\/hawk-unblock.sh/D' /var/spool/cron/root
-warn_err $? "Could not remove hawk-unblock.sh from crontab"
-
-# restart crond
-if [ -d /usr/local/1h/lib/guardian/svcstop ]; then
-	touch /usr/local/1h/lib/guardian/svcstop/crond
+	sed -i '/usr\/local\/1h\/bin\/hawk-unblock.sh/D' /var/spool/cron/root
+	warn_err $? "Could not remove hawk-unblock.sh from crontab"
 fi
 
-if ( ! /etc/init.d/crond restart ); then
-	echo "/etc/init.d/crond restart failed"
-	exit 1
-fi
+if [ -x /etc/init.d/crond ]; then
+	# restart crond
+	if [ -d /usr/local/1h/lib/guardian/svcstop ]; then
+		touch /usr/local/1h/lib/guardian/svcstop/crond
+	fi
 
-rm -f /usr/local/1h/lib/guardian/svcstop/crond
+	if ( ! /etc/init.d/crond restart ); then
+		echo "/etc/init.d/crond restart failed"
+		exit 1
+	fi
+
+	rm -f /usr/local/1h/lib/guardian/svcstop/crond
+fi
